// src/generated/responses.ts
export interface UserLocationClass {
  name: string;
}
// ...
export interface ImagesClass {
  items: AvatarImageElement[];
}
// ...
export interface GetPartnerArtistConcertsData {
  artistUnion: GetPartnerArtistConcertsDataArtistUnion;
  nearby: Nearby;
  concerts: GetPartnerArtistConcertsDataConcerts;
}

export interface GetPartnerArtistConcertsDataArtistUnion {
  headerImage: ThumbnailImageClass;
  profile: UserLocationClass;
  uri: string;
  visuals: FluffyVisuals;
}

export interface ThumbnailImageClass {
  data: BackgroundImageClass;
}

export interface FluffyVisuals {
  avatarImage: AvatarImage;
}

export interface AvatarImage {
  extractedColors: FluffyExtractedColors;
}

export interface FluffyExtractedColors {
  colorDark: Color;
}

export interface GetPartnerArtistConcertsDataConcerts {
  concerts: ImagesClass;
}

export interface Nearby {
  concerts: ImagesClass;
  locationName: string;
}
// ...
export interface GetPartnerConcertData {
  concert: Concert;
}

export interface Concert {
  __typename: string;
  message: string;
}

export interface GetPartnerConcertLocationsData {
  concertLocations: ConcertLocations;
}

export interface ConcertLocations {
  items: ConcertLocationsItem[];
}

export interface ConcertLocationsItem {
  country: Market;
  fullName: string;
  geoHash: string;
  geonameId: string;
  name: string;
}
// ...
export interface GetPartnerSearchConcertArtistsData {
  artist: GetPartnerSearchConcertArtistsDataArtist;
  concerts: GetPartnerArtistConcertsData;
  detailsEnriched: null;
  error: null;
}

export interface GetPartnerSearchConcertArtistsDataArtist {
  id: string;
  uri: string;
  name: string;
  imageUrl: string;
  followers: null;
  genres: null;
}